requirements.txt
Flask
git+https://github.com/mattupstate/flask-principal#egg=Flask-Principal
Flask-Script
Flask-SQLAlchemy
Flask-WTF
WTForms-Alchemy
webargs
